CDE-302 : Refactor : Déplacement objets dans package model
