+++ src/main/java/fr/abes/bestppn/kafka/TopicProducer.java
@@ -3,8 +3,8 @@ package fr.abes.bestppn.kafka;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import fr.abes.LigneKbartConnect;
 import fr.abes.LigneKbartImprime;
-import fr.abes.bestppn.dto.kafka.LigneKbartDto;
-import fr.abes.bestppn.entity.bacon.ProviderPackage;
+import fr.abes.bestppn.model.dto.kafka.LigneKbartDto;
+import fr.abes.bestppn.model.entity.bacon.ProviderPackage;
 import fr.abes.bestppn.exception.BestPpnException;
 import fr.abes.bestppn.utils.UtilsMapper;
 import jakarta.annotation.PostConstruct;
