+++ src/main/java/fr/abes/bestppn/model/entity/basexml/notice/Controlfield.java
@@ -1,4 +1,4 @@
-package fr.abes.bestppn.entity.basexml.notice;
+package fr.abes.bestppn.model.entity.basexml.notice;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlText;
