+++ src/main/java/fr/abes/bestppn/repository/bacon/LigneKbartRepository.java
@@ -1,7 +1,7 @@
 package fr.abes.bestppn.repository.bacon;
 
 import fr.abes.bestppn.configuration.BaconDbConfiguration;
-import fr.abes.bestppn.entity.bacon.LigneKbart;
+import fr.abes.bestppn.model.entity.bacon.LigneKbart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
