+++ src/main/java/fr/abes/bestppn/repository/basexml/NoticesBibioRepository.java
@@ -1,6 +1,6 @@
 package fr.abes.bestppn.repository.basexml;
 
-import fr.abes.bestppn.entity.basexml.NoticesBibio;
+import fr.abes.bestppn.model.entity.basexml.NoticesBibio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
