+++ src/main/java/fr/abes/bestppn/service/LogFileService.java
@@ -1,6 +1,6 @@
 package fr.abes.bestppn.service;
 
-import fr.abes.bestppn.entity.ExecutionReport;
+import fr.abes.bestppn.model.entity.ExecutionReport;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
