+++ src/main/java/fr/abes/bestppn/service/ProviderService.java
@@ -1,7 +1,7 @@
 package fr.abes.bestppn.service;
 
-import fr.abes.bestppn.entity.bacon.Provider;
-import fr.abes.bestppn.entity.bacon.ProviderPackage;
+import fr.abes.bestppn.model.entity.bacon.Provider;
+import fr.abes.bestppn.model.entity.bacon.ProviderPackage;
 import fr.abes.bestppn.exception.IllegalDateException;
 import fr.abes.bestppn.exception.IllegalPackageException;
 import fr.abes.bestppn.repository.bacon.LigneKbartRepository;
