+++ src/main/java/fr/abes/bestppn/service/WsService.java
@@ -2,12 +2,10 @@ package fr.abes.bestppn.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import fr.abes.bestppn.dto.wscall.ResultDat2PpnWebDto;
-import fr.abes.bestppn.dto.wscall.ResultWsSudocDto;
-import fr.abes.bestppn.dto.wscall.SearchDatWebDto;
-import fr.abes.bestppn.exception.BestPpnException;
+import fr.abes.bestppn.model.dto.wscall.ResultDat2PpnWebDto;
+import fr.abes.bestppn.model.dto.wscall.ResultWsSudocDto;
+import fr.abes.bestppn.model.dto.wscall.SearchDatWebDto;
 import fr.abes.bestppn.exception.IllegalDoiException;
-import fr.abes.bestppn.utils.ExecutionTime;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
